requirements.txt
# ── Orquestación ───────────────────────────────────────────────────
crewai>=0.80.0
crewai-tools>=0.14.0

# ── LLM providers ──────────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-aws>=0.2.0
langchain-openai>=0.2.0

# ── Search APIs ────────────────────────────────────────────────────
tavily-python>=0.5.0
duckduckgo-search>=6.0.0

# ── HTTP / Scraping ────────────────────────────────────────────────
httpx>=0.27.0
beautifulsoup4>=4.12.0
playwright>=1.44.0
lxml>=5.0.0

# ── Fuzzy dedup ────────────────────────────────────────────────────
rapidfuzz>=3.0.0
unidecode>=1.3.0

# ── Output ─────────────────────────────────────────────────────────
openpyxl>=3.1.0

# ── Config / validation ────────────────────────────────────────────
pydantic>=2.6.0
pydantic-settings>=2.2.0
pyyaml>=6.0.0

# ── CLI / Console ──────────────────────────────────────────────────
rich>=13.0.0

# ── Testing ────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0

# ── AWS SDK (Bedrock) ──────────────────────────────────────────────
boto3>=1.34.0
botocore>=1.34.0
